+++ src/com/Square9/AndroidMapsV2Test/OnDialogDoneListener.java
@@ -8,6 +8,6 @@ package com.Square9.AndroidMapsV2Test;
 public interface OnDialogDoneListener
 {
     public void onDialogDone(String tag, boolean cancelled, CharSequence message);
-    public void onDialogDone(String tag, boolean cancelled, String ln, float color, int lw);
+    public void onDialogDone(String tag, boolean cancelled, String ln, int color, int lw);
 
 }
